+++ src/Query/CreateTable.php
@@ -7,7 +7,6 @@ use Formal\AccessLayer\{
     Query,
     Query\Constraint\PrimaryKey,
     Query\Constraint\ForeignKey,
-    Row,
     Table\Name,
     Table\Column,
 };